+++ NEWS.md
@@ -1,3 +1,5 @@
+# valr (development version)
+
 # valr 0.6.3
 
 * Update to prepare for readr 2.0.0