+++ NEWS.md
@@ -1,6 +1,7 @@
 # valr (development version)
 
-* Fixed intron score numbering error in `create_introns` (#377 @sheridar) 
+* Fixed intron score numbering error in `create_introns` (#377 @sheridar)  
+* Fixed bug in handling of list inputs for `bed_intersect()`(#380 @sheridar)  
 
 # valr 0.6.3